+++ R/womblR-package.R
@@ -7,9 +7,12 @@
 #' at each time point through a conditional autoregressive (CAR) prior. Temporal
 #' correlation is introduced through a hierarchical structure and can be specified as
 #' exponential or first-order autoregressive. Full details of the the package can be found
-#' in the accompanying vignette.
+#' in the accompanying vignette. Furthermore, the details of the package can be found in
+#' "Diagnosing Glaucoma Progression with Visual Field Data Using a Spatiotemporal Boundary
+#' Detection Method", by Berchuck et al (2018), <arXiv:1805.11636>. The paper is in press
+#' at the Journal of the American Statistical Association.
 #'
-#' @author Samuel I. Berchuck \email{berchuck@ad.unc.edu}
+#' @author Samuel I. Berchuck \email{sib2@duke.edu}
 #'
 #' @name womblR
 #' @docType package